week2/pyproject.toml
[tool.poetry]
name = "week2"
version = "0.1.0"
description = ""
authors = ["raccoony <raccoonyy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
requests = "^2.28.0"
SQLAlchemy = "^1.4.37"
python-multipart = "^0.0.5"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^3.2.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-watch = "^4.2.0"
parameterized = "^0.8.1"
freezegun = "^1.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
